Completed Insertion Query
Added the insertion query to the php code, and caused the page to automatically redirect back to the book table once book entry is finished. Nesrin may want to double check that the capitalization and such is right on the attribute names in the query, because I'm currently working without access to the database.